+++ skeletons/VisibleString.c
@@ -51,10 +51,10 @@ VisibleString_constraint(asn_TYPE_descriptor_t *td, const void *sptr,
 		for(; buf < end; buf++) {
 			if(*buf < 0x20 || *buf > 0x7e) {
 				_ASN_ERRLOG(app_errlog, app_key,
-					"%s: value byte %d (%d) "
+					"%s: value byte %ld (%d) "
 					"not in VisibleString alphabet (%s:%d)",
 					td->name,
-					(buf - st->buf) + 1,
+					(long)((buf - st->buf) + 1),
 					*buf,
 					__FILE__, __LINE__);
 				return -1;
